Refactoring tracker and classifiers. Moving live_feed as a logger.
