+++ libs/area_threading.py
@@ -36,25 +36,29 @@ class AreaThread(Thread):
         self.area = area
 
     def run(self):
-        self.engine = AreaEngine(self.config, self.area)
-        restarts = 0
-        max_restarts = int(self.config.get_section_dict("App")["MaxThreadRestarts"])
-        while True:
-            try:
-                last_restart_time = datetime.now()
-                self.engine.process_area()
-            except Exception as e:
-                logging.error(e, exc_info=True)
-                logging.info(f"Exception processing area {self.area['name']}")
-                if (datetime.now() - last_restart_time).total_seconds() > 60:
-                    # If the last restart was previous than 1 minute ago, restart the counter.
-                    restarts = 0
-                if restarts == max_restarts:
-                    raise e
-                # Sleep the thread for 5 seconds and try to process the area again
-                time.sleep(5)
-                logging.info("Restarting the area processing")
-                restarts += 1
+        try:
+            self.engine = AreaEngine(self.config, self.area)
+            restarts = 0
+            max_restarts = int(self.config.get_section_dict("App")["MaxThreadRestarts"])
+            while True:
+                try:
+                    last_restart_time = datetime.now()
+                    self.engine.process_area()
+                except Exception as e:
+                    logging.error(e, exc_info=True)
+                    logging.info(f"Exception processing area {self.area['name']}")
+                    if (datetime.now() - last_restart_time).total_seconds() > 60:
+                        # If the last restart was previous than 1 minute ago, restart the counter.
+                        restarts = 0
+                    if restarts == max_restarts:
+                        raise e
+                    # Sleep the thread for 5 seconds and try to process the area again
+                    time.sleep(5)
+                    logging.info("Restarting the area processing")
+                    restarts += 1
+        except Exception as e:
+            logging.error(e, exc_info=True)
+            raise e
 
     def stop(self):
         self.engine.stop_process_area()
